IMN 380 - Move Agreement converter into models package (#373)
