+++ packages/agreement-process/test/cloneAgreement.test.ts
@@ -24,6 +24,7 @@ import {
   TenantId,
   agreementState,
   generateId,
+  toAgreementV2,
   unsafeBrandId,
 } from "pagopa-interop-models";
 import { FileManagerError, genericLogger } from "pagopa-interop-commons";
@@ -31,7 +32,6 @@ import {
   agreementClonableStates,
   agreementCloningConflictingStates,
 } from "../src/model/domain/validators.js";
-import { toAgreementV2 } from "../src/model/domain/toEvent.js";
 import { config } from "../src/utilities/config.js";
 import {
   agreementAlreadyExists,
