+++ packages/agreement-process/test/suspendAgreement.test.ts
@@ -31,6 +31,7 @@ import {
   VerifiedTenantAttribute,
   agreementState,
   generateId,
+  toAgreementV2,
 } from "pagopa-interop-models";
 import { genericLogger } from "pagopa-interop-commons";
 import { agreementSuspendableStates } from "../src/model/domain/validators.js";
@@ -43,7 +44,6 @@ import {
   operationNotAllowed,
   tenantNotFound,
 } from "../src/model/domain/errors.js";
-import { toAgreementV2 } from "../src/model/domain/toEvent.js";
 import {
   addOneAgreement,
   addOneEService,
