+++ packages/agreement-process/test/upgradeAgreement.test.ts
@@ -33,6 +33,7 @@ import {
   agreementState,
   descriptorState,
   generateId,
+  toAgreementV2,
   unsafeBrandId,
 } from "pagopa-interop-models";
 import { afterAll, beforeAll, describe, expect, it, vi } from "vitest";
@@ -49,7 +50,6 @@ import {
   tenantNotFound,
   unexpectedVersionFormat,
 } from "../src/model/domain/errors.js";
-import { toAgreementV2 } from "../src/model/domain/toEvent.js";
 import { agreementUpgradableStates } from "../src/model/domain/validators.js";
 import { config } from "../src/utilities/config.js";
 import {
