+++ packages/models/src/index.ts
@@ -4,9 +4,10 @@ export * from "./events/events.js";
 // Entities, events, converters
 export * from "./agreement/agreement.js";
 export * from "./agreement/agreementEvents.js";
-export * from "./agreement/agreementReadModelAdapter.js";
 export * from "./agreement/protobufConverterFromV1.js";
 export * from "./agreement/protobufConverterFromV2.js";
+export * from "./agreement/protobufConverterToV2.js";
+export * from "./agreement/agreementReadModelAdapter.js";
 
 export * from "./attribute/attribute.js";
 export * from "./attribute/attributeReadModelAdapter.js";
